--- packages/calendar/styles/calendar.css
@@ -1,1 +1,0 @@
-@import url('./days.css');
