--- packages/calendar/styles/days.css
@@ -1,91 +1,0 @@
-#days .fc .fc-toolbar-title {
-    font-family: var(--fuente-tokko);
-    font-style: normal;
-    font-weight: 700;
-    font-size: 13px;
-    line-height: 18px;
-    color: var(--secondary);
-}
-
-#days .fc .fc-day-today:not(.fc-popover):not(.fc-col-header-cell) {
-    background-color: transparent;
-}
-
-#days .fc-theme-standard td,
-#days .fc-theme-standard th {
-    border: none;
-    border-top: 1px solid var(--neutralGray7);
-}
-
-#days .fc .fc-cell-shaded,
-#days .fc .fc-day-disabled {
-    background: transparent;
-}
-
-#days .fc .fc-timegrid-axis-cushion,
-#days .fc .fc-timegrid-slot-label-cushion {
-    font-family: var(--fuente-tokko);
-    font-style: normal;
-    font-weight: 400;
-    font-size: 12px;
-    line-height: 16px;
-    color: var(--neutralGray2);
-}
-
-#days .fc .fc-button-primary {
-    box-shadow: none !important;
-}
-
-#days .fc .fc-button-primary {
-    background-color: transparent !important;
-    border: none !important;
-    color: var(--neutralGray2) !important;
-}
-
-#days .fc-timegrid-event .fc-event-main {
-    padding: 3px 6px !important;
-}
-
-#days .fc-timegrid-event-harness-inset .fc-timegrid-event,
-#days .fc-timegrid-event.fc-event-mirror,
-#days .fc-timegrid-more-link {
-    border-radius: 8px;
-    border: none;
-    background-color: var(--neutralGray3);
-}
-
-#days .fc-timegrid-event .fc-event-time,
-#days .fc-v-event .fc-event-title-container {
-    font-family: var(--fuente-tokko);
-    font-size: 11px;
-    line-height: 13px;
-}
-
-#days .fc-v-event .fc-event-title-container {
-    font-weight: 700;
-}
-
-#days .fc-scrollgrid-section:first-child,
-#days .fc-scrollgrid-section:nth-child(2) {
-    display: none;
-}
-
-#days .fc-theme-standard .fc-scrollgrid {
-    border: none;
-}
-
-#days .fc-scroller-liquid-absolute::-webkit-scrollbar {
-    width: 2px;
-}
-
-#days .fc-scroller-liquid-absolute::-webkit-scrollbar-track {
-    background-color: white;
-}
-
-#days .fc-scroller-liquid-absolute::-webkit-scrollbar-thumb {
-    background-color: #6F838D
-}
-
-#days .fc .fc-button {
-    padding: 5px 10px;
-}
